requirements.txt
distro==1.9.0
et-xmlfile==1.1.0
JPype1==1.5.0
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==15.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
tabula-py==2.9.0
tzdata==2023.4
